+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Prop `plain` to `ListboxButton` to remove borders.
+- Prop `variation` to `ListboxButton` to control component appearance.
 
 ## [0.3.0] - 2020-05-28
 ### Added